BE-1345: handle exception when website is forced to use document mode 8 - include jQuery and jQueryUI in all.js for Chrome, IE, Safari (server_phases.py) - prevent load jQuery for Chrome, IE, Safari (manifest.json) - wrap all.js in a function scope to prevent variable name conflict and stop execute scripts if IE8 (server_phases.py) - wrap (forge.html) to stop load script if IE8 - FF: do not wrap jQuery libraries (customer_tasks.py) - handle exception console.info for IE (api-ie.js)

+++ generate/generate/customer_tasks.py
@@ -516,13 +516,14 @@ def wrap_activations(build, location):
 				for script in activation['scripts']:
 					tmp_file = uuid.uuid4().hex
 					filename = location+script[3:]
-					build.log.debug("wrapping activation {filename}".format(**locals()))
-					in_file_contents = read_file_as_str(filename)
-					in_file_contents = '// firefox complains when the first line is an if statement\n' + 'if (forge._disableFrames === undefined || window.location == window.parent.location) {\n' + in_file_contents + '\n}';
-					with codecs.open(tmp_file, 'w', encoding='utf8') as out_file:
-						out_file.write(in_file_contents)
-					os.remove(filename)
-					shutil.move(tmp_file, filename)
+					if not filename.endswith(("jquery-2.2.4.js", "jquery-ui-1.10.4.custom.js")):
+						build.log.debug("wrapping activation {filename}".format(**locals()))
+						in_file_contents = read_file_as_str(filename)
+						in_file_contents = '// firefox complains when the first line is an if statement\n' + 'if (forge._disableFrames === undefined || window.location == window.parent.location) {\n' + in_file_contents + '\n}';
+						with codecs.open(tmp_file, 'w', encoding='utf8') as out_file:
+							out_file.write(in_file_contents)
+						os.remove(filename)
+						shutil.move(tmp_file, filename)
 		
 @task
 def populate_icons(build, platform, icon_list):
